refactor: extract the shared registry, source and OCI layer into pkg/registry/component
A module and an addon are fetched from the same kinds of place, so they were
reading them with two copies of the same code. This moves the shared half into
pkg/registry/component: the registry model and its ConfigMap-backed store, the
source types and their credential handling, the git/gitee/gitlab/OSS/local/memory
readers, and the OCI chart transport.

pkg/addon keeps what is actually addon-specific -- interpreting fetched files as
an addon (GetUIData, ListUIData, GetInstallPackage, ListAddonInfo,
ClassifyItemByPattern), the versioned-registry backends, and catalog listing --
and reaches the shared half through component_alias.go, so existing spellings
such as addon.Registry keep working and no caller outside this package changes.
Functions that used to be methods on Registry are now free functions taking one,
because a package cannot define methods on a type it does not own.

The OCI transport moves rather than being duplicated, which is the point: the
addon path and the module path had grown separate copies that disagreed. One
parsed only oci:// URLs while the other also accepted http:// and scheme-less
registry hosts, so the same registry record could resolve differently depending
on which path reached it. The surviving implementation keeps both behaviours
that mattered: the wider URL handling, and the client cache, per-call timeout
and context-aware AwaitOCICall that let a cancelled caller go without leaking
the connection.
