+++ pkg/addon/addon_test.go
@@ -49,6 +49,7 @@ import (
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/v1beta1"
 	"github.com/oam-dev/kubevela/apis/types"
 	"github.com/oam-dev/kubevela/pkg/oam"
+	"github.com/oam-dev/kubevela/pkg/registry/component"
 	addonutil "github.com/oam-dev/kubevela/pkg/utils/addon"
 	httpoption "github.com/oam-dev/kubevela/pkg/utils/common"
 	version2 "github.com/oam-dev/kubevela/version"
@@ -703,10 +704,7 @@ func TestGetPatternFromItem(t *testing.T) {
 	gitItemType := FileType
 	gitItemPath := "addons/terraform/resources/parameter.cue"
 
-	viewOSSR := localReader{
-		dir:  "./testdata/test-view",
-		name: "test-view",
-	}
+	viewOSSR := component.NewLocalReader("./testdata/test-view", "test-view")
 	viewPath := filepath.Join("./testdata/test-view/views/pod-view.cue", "pod-view.cue")
 
 	testCases := []struct {
@@ -717,12 +715,8 @@ func TestGetPatternFromItem(t *testing.T) {
 		r           AsyncReader
 	}{
 		{
-			caseName: "OSS case",
-			item: OSSItem{
-				tp:   FileType,
-				path: "terraform/resources/parameter.cue",
-				name: "parameter.cue",
-			},
+			caseName:    "OSS case",
+			item:        component.NewOSSItem(FileType, "terraform/resources/parameter.cue", "parameter.cue"),
 			root:        "terraform",
 			meetPattern: "resources/parameter.cue",
 			r:           ossR,
@@ -735,12 +729,8 @@ func TestGetPatternFromItem(t *testing.T) {
 			r:           gitR,
 		},
 		{
-			caseName: "views case",
-			item: OSSItem{
-				tp:   FileType,
-				path: viewPath,
-				name: "pod-view.cue",
-			},
+			caseName:    "views case",
+			item:        component.NewOSSItem(FileType, viewPath, "pod-view.cue"),
 			root:        "test-view",
 			meetPattern: "views",
 			r:           viewOSSR,
@@ -998,24 +988,13 @@ var testUnmarshalToContent3 = `
 ]`
 var testUnmarshalToContent4 = ``
 
-func TestUnmarshalToContent(t *testing.T) {
-	_, _, err1 := unmarshalToContent([]byte(testUnmarshalToContent1))
-	assert.NoError(t, err1)
-	_, _, err2 := unmarshalToContent([]byte(testUnmarshalToContent2))
-	assert.NoError(t, err2)
-	_, _, err3 := unmarshalToContent([]byte(testUnmarshalToContent3))
-	assert.Error(t, err3, "unmarshalling failed for both file and directory content: invalid character '}' looking for beginnin")
-	_, _, err4 := unmarshalToContent([]byte(testUnmarshalToContent4))
-	assert.Error(t, err4, "unmarshalling failed for both file and directory content: unexpected end of JSON input and unexpecte")
-}
-
 // Test readResFile, only accept .cue and .yaml/.yml
 func TestReadResFile(t *testing.T) {
 
 	// setup test data
 	testAddonName := "example"
 	testAddonDir := fmt.Sprintf("./testdata/%s", testAddonName)
-	reader := localReader{dir: testAddonDir, name: testAddonName}
+	reader := component.NewLocalReader(testAddonDir, testAddonName)
 	metas, err := reader.ListAddonMeta()
 	testAddonMeta := metas[testAddonName]
 	assert.NoError(t, err)
@@ -1039,7 +1018,7 @@ func TestReadDefFile(t *testing.T) {
 	// setup test data
 	testAddonName := "example"
 	testAddonDir := fmt.Sprintf("./testdata/%s", testAddonName)
-	reader := localReader{dir: testAddonDir, name: testAddonName}
+	reader := component.NewLocalReader(testAddonDir, testAddonName)
 	metas, err := reader.ListAddonMeta()
 	testAddonMeta := metas[testAddonName]
 	assert.NoError(t, err)
@@ -1066,7 +1045,7 @@ func TestReadViewFile(t *testing.T) {
 	// setup test data
 	testAddonName := "test-view"
 	testAddonDir := fmt.Sprintf("./testdata/%s", testAddonName)
-	reader := localReader{dir: testAddonDir, name: testAddonName}
+	reader := component.NewLocalReader(testAddonDir, testAddonName)
 	metas, err := reader.ListAddonMeta()
 	testAddonMeta := metas[testAddonName]
 	assert.NoError(t, err)
